+++ example/index.tsx
@@ -1,10 +1,8 @@
-import * as React from 'react'
-import { render } from 'react-dom'
+import { createRoot } from 'react-dom/client'
 import { AppInitializer } from './App'
 
-render(
-    <React.StrictMode>
-        <AppInitializer />
-    </React.StrictMode>,
-    document.getElementById('root')
-)
+const container = document.getElementById('root')
+// eslint-disable-next-line @typescript-eslint/no-non-null-assertion
+const root = createRoot(container!)
+
+root.render(<AppInitializer />)