+++ test/state/state.test.ts
@@ -1,4 +1,5 @@
-import { of, Observable } from 'rxjs'
+import type { Observable } from 'rxjs'
+import { of } from 'rxjs'
 import { TestScheduler } from 'rxjs/testing'
 import * as MS from '../../src/state/index'
 
@@ -27,11 +28,11 @@ describe('state', () => {
 
         const task = MS.toTask(store)
 
-        store.state$.next({ name: 'hey' })
+        store.setState({ name: 'hey' })
         const stateUpdatedHey = await task()
         expect(stateUpdatedHey.name).toBe('hey')
 
-        store.state$.next({ name: 'ho' })
+        store.setState({ name: 'ho' })
         const stateUpdatedResultHo = await task()
         expect(stateUpdatedResultHo.name).toBe('ho')
     })